feat: extract shim implementation into an own module for re-use
