+++ docs/observe.md
@@ -1,0 +1,131 @@
+# twind/observe
+
+> Allows to copy-paste tailwind examples. This feature can be used together with your favorite framework without any additional setup.
+
+The `twind/observe` modules allows to use the `class` attribute for tailwind rules. If such a rule is detected the corresponding CSS rule is created and injected into the stylesheet. _No need for `tw`_ but it can be used on the same elements as well.
+
+> This is meant for advanced use cases. Most of the time you may want to use [twind/shim](./installation.md#twindshim).
+
+<details><summary>Table Of Contents (Click To Expand)</summary>
+
+<!-- START doctoc generated TOC please keep comment here to allow auto update -->
+<!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
+
+- [Usage](#usage)
+- [Customization](#customization)
+- [API](#api)
+- [Example](#example)
+- [Implementation Details](#implementation-details)
+
+<!-- END doctoc generated TOC please keep comment here to allow auto update -->
+</details>
+
+## Usage
+
+```js
+import { observe } from 'twind/observe'
+
+observe(document.body)
+
+document.body.innerHTML = `
+  <main class="h-screen bg-purple-400 flex items-center justify-center">
+    <h1 class="font-bold text(center 5xl white sm:gray-800 md:pink-700)">
+      This is Twind!
+    </h1>
+  </main>
+`
+```
+
+> [live and interactive shim demo](https://esm.codes/#aW1wb3J0IHsgb2JzZXJ2ZSB9IGZyb20gJ3R3aW5kL29ic2VydmUnCgpvYnNlcnZlKGRvY3VtZW50LmJvZHkpCgpkb2N1bWVudC5ib2R5LmlubmVySFRNTCA9IGAKICA8bWFpbiBjbGFzcz0iaC1zY3JlZW4gYmctcHVycGxlLTQwMCBmbGV4IGl0ZW1zLWNlbnRlciBqdXN0aWZ5LWNlbnRlciI+CiAgICA8aDEgY2xhc3M9ImZvbnQtYm9sZCB0ZXh0KGNlbnRlciA1eGwgd2hpdGUgc206Z3JheS04MDAgbWQ6cGluay03MDApIj4KICAgICAgVGhpcyBpcyBUd2luZCEKICAgIDwvaDE+CiAgPC9tYWluPgpg)
+
+All twind syntax features like [grouping](./grouping.md) are supported.
+
+> If you want to simplify the instantiation and automatically observe take look at [twind/shim](./installation.md#twindshim).
+
+## Customization
+
+`twind/observe` uses the default/global `tw` instance if not configured otherwise. You can provide a custom instance in several ways:
+
+```js
+import { create } from 'twind'
+import { observe, createObserver } from 'twind/observe'
+
+// Create a custom instance
+const instance = create(/* ... */)
+
+// 1. As second parameter
+observe(document.body, instance)
+
+// 2. As this context
+observe.call(document.body)
+observe.bind(instance)(document.body)
+
+// 3. Use the factory
+createObserver(instance).observe(document.body)
+```
+
+## API
+
+```js
+import { createObserver, observe } from 'twind/observe'
+
+const observer = createObserver(/* custom instance */)
+// Or to start observing an element right away
+// const observer = observe(document.body)
+
+// Start observing a node
+observer.observe(node)
+
+// Stop observing
+observer.disconnect(node)
+```
+
+## Example
+
+> This example show how a custom observer instance can be used to shim a web component.
+
+```js
+import { LitElement, html } from 'lit-element'
+import { create, cssomSheet } from 'twind'
+import { createObserver } from 'twind/observe'
+
+// 1. Create separate CSSStyleSheet
+const sheet = cssomSheet({ target: new CSSStyleSheet() })
+
+// 2. Use that to create an own twind instance
+const instance = create({ sheet })
+
+class TwindElement extends LitElement {
+  // 3. Apply the same style to each instance of this component
+  static styles = [sheet.target]
+
+  // 4. Start observing class attributes changes
+  connectedCallback() {
+    this._observer = createObserver(instance).observe(this.renderRoot)
+  }
+
+  // 5. Stop observing class attributes changes
+  disconnectedCallback() {
+    this._observer.disconnect()
+  }
+
+  render() {
+    // 5. Use tailwind rules in class attributes
+    return html`
+      <main class="h-screen bg-purple-400 flex items-center justify-center">
+        <h1 class="font-bold text(center 5xl white sm:gray-800 md:pink-700)">This is Twind!</h1>
+      </main>
+    `
+  }
+}
+
+customElements.define('twind-element', TwindElement)
+
+document.body.innerHTML = '<twind-element></twind-element>'
+```
+
+> [live and interactive demo](https://esm.codes/#aW1wb3J0IHsgTGl0RWxlbWVudCwgaHRtbCB9IGZyb20gJ2h0dHBzOi8vY2RuLnNreXBhY2suZGV2L2xpdC1lbGVtZW50JwppbXBvcnQgeyBjcmVhdGUsIGNzc29tU2hlZXQgfSBmcm9tICdodHRwczovL2Nkbi5za3lwYWNrLmRldi90d2luZCcKaW1wb3J0IHsgY3JlYXRlT2JzZXJ2ZXIgfSBmcm9tICdodHRwczovL2Nkbi5za3lwYWNrLmRldi90d2luZC9vYnNlcnZlJwoKY29uc3Qgc2hlZXQgPSBjc3NvbVNoZWV0KHsgdGFyZ2V0OiBuZXcgQ1NTU3R5bGVTaGVldCgpIH0pCgpjb25zdCB7IHR3IH0gPSBjcmVhdGUoeyBzaGVldCB9KQoKY2xhc3MgVHdpbmRFbGVtZW50IGV4dGVuZHMgTGl0RWxlbWVudCB7CiAgY3JlYXRlUmVuZGVyUm9vdCgpIHsKICAgIGNvbnN0IHNoYWRvdyA9IHN1cGVyLmNyZWF0ZVJlbmRlclJvb3QoKQogICAgc2hhZG93LmFkb3B0ZWRTdHlsZVNoZWV0cyA9IFtzaGVldC50YXJnZXRdCiAgICByZXR1cm4gc2hhZG93CiAgfQoKICBjb25uZWN0ZWRDYWxsYmFjaygpIHsKICAgIHRoaXMuX29ic2VydmVyID0gY3JlYXRlT2JzZXJ2ZXIoaW5zdGFuY2UpLm9ic2VydmUodGhpcy5yZW5kZXJSb290KQogIH0KCiAgZGlzY29ubmVjdGVkQ2FsbGJhY2soKSB7CiAgICB0aGlzLl9vYnNlcnZlci5kaXNjb25uZWN0KCkKICB9CgogIHJlbmRlcigpIHsKICAgIHJldHVybiBodG1sYAogICAgICA8bWFpbiBjbGFzcz0iaC1zY3JlZW4gYmctcHVycGxlLTQwMCBmbGV4IGl0ZW1zLWNlbnRlciBqdXN0aWZ5LWNlbnRlciI+CiAgICAgICAgPGgxIGNsYXNzPSJmb250LWJvbGQgdGV4dChjZW50ZXIgNXhsIHdoaXRlIHNtOmdyYXktODAwIG1kOnBpbmstNzAwKSI+CiAgICAgICAgICBUaGlzIGlzIFR3aW5kIQogICAgICAgIDwvaDE+CiAgICAgIDwvbWFpbj4KICAgIGAKICB9Cn0KCmN1c3RvbUVsZW1lbnRzLmRlZmluZSgndHdpbmQtZWxlbWVudCcsIFR3aW5kRWxlbWVudCk7Cgpkb2N1bWVudC5ib2R5LmlubmVySFRNTCA9ICc8dHdpbmQtZWxlbWVudD48L3R3aW5kLWVsZW1lbnQ+Jwo=)
+
+## Implementation Details
+
+This uses a [MutationObserver](https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver) to detect changed class attributes or added DOM nodes. On detection the class attribute is parsed and translated by twind to inject the required classes into the stylesheet and the class attribute is updated to reflect the added CSS class names that may have been hashed.
