+++ src/shim/index.ts
@@ -1,66 +1,12 @@
 import type { Configuration } from '../types'
-import { tw, setup as setupTW } from '../index'
+import type { TwindObserver } from '../observe'
+
+import { setup as setupTW } from '../index'
+import { createObserver } from '../observe/index'
 
 export interface ShimConfiguration extends Configuration {
   target?: HTMLElement
 }
-
-interface NodeList {
-  readonly length: number
-
-  [index: number]: Node
-}
-
-interface MutationRecord {
-  readonly addedNodes: NodeList
-  readonly target: Node
-}
-
-const rulesToClassCache = new Map<string, string>()
-
-const handleMutation = ({ target, addedNodes }: MutationRecord): void => {
-  // Not using target.classList.value (not supported in all browsers) or target.class (this is an SVGAnimatedString for svg)
-  const rules = (target as Element).getAttribute?.('class')
-
-  if (rules) {
-    let className = rulesToClassCache.get(rules)
-
-    if (!className) {
-      className = tw(rules)
-
-      // Remember the generated class name
-      rulesToClassCache.set(rules, className)
-
-      // Ensure the cache does not grow unlimited
-      if (rulesToClassCache.size > 10000) {
-        rulesToClassCache.delete(rulesToClassCache.keys().next().value)
-      }
-    }
-
-    if (rules !== className) {
-      // Not using `target.className = ...` as that is read-only for SVGElements
-      // eslint-disable-next-line @typescript-eslint/no-extra-semi
-      ;(target as Element).setAttribute('class', className)
-    }
-  }
-
-  for (let index = addedNodes.length; index--; ) {
-    const node = addedNodes[index]
-
-    handleMutations([
-      {
-        target: node,
-        addedNodes: (node as Element).children || [],
-      },
-    ])
-  }
-}
-
-const handleMutations = (mutations: MutationRecord[]): void => mutations.forEach(handleMutation)
-
-const observer = new MutationObserver(handleMutations)
-
-export const stop = (): void => observer.disconnect()
 
 const onload = () => {
   const script = document.querySelector('script[type="twind-config"]')
@@ -78,29 +24,29 @@ if (document.readyState === 'loading') {
   var timeoutRef = setTimeout(onload)
 }
 
+const observer = createObserver()
+
+export const disconnect = (): void => {
+  // Removing the callbacks ensures that the setup is called only once
+  // either programmatically from userland or by DOMContentLoaded/setTimeout
+  removeEventListener('DOMContentLoaded', onload)
+  clearTimeout(timeoutRef)
+
+  observer.disconnect()
+}
+
 export const setup = ({
   target = document.documentElement,
   ...config
 }: ShimConfiguration = {}): void => {
-  // Removing the callbacks ensures that the setup is called only once
-  // either by programmatically from userland or by DOMContentLoaded/timeout
-  removeEventListener('DOMContentLoaded', onload)
-  clearTimeout(timeoutRef)
-
   if (Object.keys(config).length) {
     setupTW(config)
   }
 
-  stop()
+  // Remove event listeners
+  disconnect()
 
-  handleMutations([{ target, addedNodes: [target] }])
+  observer.observe(target)
 
   target.hidden = false
-
-  observer.observe(target, {
-    attributes: true,
-    attributeFilter: ['class'],
-    subtree: true,
-    childList: true,
-  })
 }
